+++ lib/kdecorationbridge.h
@@ -98,11 +98,11 @@ class KWIN_EXPORT KDecorationBridgeUnstable
         // Window tabbing
         virtual bool isClientGroupActive() = 0;
         virtual QList< ClientGroupItem > clientGroupItems() const = 0;
-        virtual int itemId( int index ) = 0;
+        virtual long itemId( int index ) = 0;
         virtual int visibleClientGroupItem() = 0;
         virtual void setVisibleClientGroupItem( int index ) = 0;
         virtual void moveItemInClientGroup( int index, int before ) = 0;
-        virtual void moveItemToClientGroup( int itemId, int before ) = 0;
+        virtual void moveItemToClientGroup( long itemId, int before ) = 0;
         virtual void removeFromClientGroup( int index, const QRect& newGeom ) = 0;
         virtual void closeClientGroupItem( int index ) = 0;
         virtual void closeAllInClientGroup() = 0;
